feat(precompiles): add Lux DEX_MATCH precompile at 0x100 (CPU ref)
Widens the precompile dispatcher's address parameter from uint8_t to
uint16_t so the standard EVM range (0x01..0x11) and the new Lux custom
range (0x100..0x1ff) both fit. The standard range is unchanged in
behaviour; numeric literals in the existing wrappers and tests promote
silently to uint16_t.

DEX_MATCH at 0x100 takes a single incoming order:

  side(1) | price(32) | qty(32) | user(20) | book_id(32)   = 117 bytes

and produces:

  filled_qty(32) | avg_price(32) | num_fills(4)            = 68 bytes

The CPU reference walks an in-memory order book keyed by book_id, sorts
the opposing side best-first, and crosses the incoming order until qty
is exhausted or no crossing levels remain. VWAP is computed via 256-bit
multiply into a 512-bit accumulator divided by total filled quantity.
Gas: 1500 base + 350 per fill, charged after matching so an OOG bails
without leaving the book in a partially-mutated state.

This commit lands the CPU path only; the GPU wiring through
liblux_accel.lux_match_orders comes next.

Parity: 133/133 (CPU_Sequential, CPU_Parallel, GPU_Metal).
Precompile tests: 30/30.

+++ lib/evm/gpu/precompiles/ecrecover_cuda.cpp
@@ -18,7 +18,7 @@
 namespace evm::gpu::precompile
 {
 extern "C" void evm_precompile_set_impl(
-    void* dispatcher, uint8_t address, Result (*fn)(std::span<const uint8_t>, uint64_t),
+    void* dispatcher, uint16_t address, Result (*fn)(std::span<const uint8_t>, uint64_t),
     int backend_id) noexcept;
 
 // CPU fallback (defined in ecrecover_cpu.cpp).
